add 'run until convergence' and fix q learning
When time_limit is -1 it now runs until convergence (i.e. the policy
doesn't change for 0.25 seconds). I also reworked the Q learning
algorithm: now it updates the table every tick, using the net score for
that tick, the state of the world the last time a decision was made, and
predicting a future state based on the current one